Milk, Soda, Coffee 클래스 파일 생성
